+++ hoomd/md/ConstraintSphereGPU.cu
@@ -99,7 +99,7 @@ void gpu_compute_constraint_sphere_forces_kernel(Scalar4* d_force,
 */
 hipError_t gpu_compute_constraint_sphere_forces(Scalar4* d_force,
                                                  Scalar* d_virial,
-                                                 const unsigned int virial_pitch,
+                                                 const size_t virial_pitch,
                                                  const unsigned int *d_group_members,
                                                  unsigned int group_size,
                                                  const unsigned int N,
